add verification email

+++ medunnes/app/src/main/java/com/medunnes/telemedicine/data/response/PasienResponse.kt
@@ -32,7 +32,7 @@ data class PasienDataItem(
 	val alamat: String,
 
 	@field:SerializedName("NIK")
-	val nik: Int,
+	val nik: Long,
 
 	@field:SerializedName("banned_at")
 	val bannedAt: Any,
